+++ src/main/java/com/smartx/tower/model/VmVolumeWhereInput.java
@@ -532,35 +532,35 @@ public class VmVolumeWhereInput {
 
   public static final String SERIALIZED_NAME_UNIQUE_LOGICAL_SIZE = "unique_logical_size";
   @SerializedName(SERIALIZED_NAME_UNIQUE_LOGICAL_SIZE)
-  private Long uniqueLogicalSize;
+  private Double uniqueLogicalSize;
 
   public static final String SERIALIZED_NAME_UNIQUE_LOGICAL_SIZE_GT = "unique_logical_size_gt";
   @SerializedName(SERIALIZED_NAME_UNIQUE_LOGICAL_SIZE_GT)
-  private Long uniqueLogicalSizeGt;
+  private Double uniqueLogicalSizeGt;
 
   public static final String SERIALIZED_NAME_UNIQUE_LOGICAL_SIZE_GTE = "unique_logical_size_gte";
   @SerializedName(SERIALIZED_NAME_UNIQUE_LOGICAL_SIZE_GTE)
-  private Long uniqueLogicalSizeGte;
+  private Double uniqueLogicalSizeGte;
 
   public static final String SERIALIZED_NAME_UNIQUE_LOGICAL_SIZE_IN = "unique_logical_size_in";
   @SerializedName(SERIALIZED_NAME_UNIQUE_LOGICAL_SIZE_IN)
-  private List<Long> uniqueLogicalSizeIn = null;
+  private List<Double> uniqueLogicalSizeIn = null;
 
   public static final String SERIALIZED_NAME_UNIQUE_LOGICAL_SIZE_LT = "unique_logical_size_lt";
   @SerializedName(SERIALIZED_NAME_UNIQUE_LOGICAL_SIZE_LT)
-  private Long uniqueLogicalSizeLt;
+  private Double uniqueLogicalSizeLt;
 
   public static final String SERIALIZED_NAME_UNIQUE_LOGICAL_SIZE_LTE = "unique_logical_size_lte";
   @SerializedName(SERIALIZED_NAME_UNIQUE_LOGICAL_SIZE_LTE)
-  private Long uniqueLogicalSizeLte;
+  private Double uniqueLogicalSizeLte;
 
   public static final String SERIALIZED_NAME_UNIQUE_LOGICAL_SIZE_NOT = "unique_logical_size_not";
   @SerializedName(SERIALIZED_NAME_UNIQUE_LOGICAL_SIZE_NOT)
-  private Long uniqueLogicalSizeNot;
+  private Double uniqueLogicalSizeNot;
 
   public static final String SERIALIZED_NAME_UNIQUE_LOGICAL_SIZE_NOT_IN = "unique_logical_size_not_in";
   @SerializedName(SERIALIZED_NAME_UNIQUE_LOGICAL_SIZE_NOT_IN)
-  private List<Long> uniqueLogicalSizeNotIn = null;
+  private List<Double> uniqueLogicalSizeNotIn = null;
 
   public static final String SERIALIZED_NAME_UNIQUE_SIZE = "unique_size";
   @SerializedName(SERIALIZED_NAME_UNIQUE_SIZE)
@@ -3723,7 +3723,7 @@ public class VmVolumeWhereInput {
   }
 
 
-  public VmVolumeWhereInput uniqueLogicalSize(Long uniqueLogicalSize) {
+  public VmVolumeWhereInput uniqueLogicalSize(Double uniqueLogicalSize) {
     
     this.uniqueLogicalSize = uniqueLogicalSize;
     return this;
@@ -3736,17 +3736,17 @@ public class VmVolumeWhereInput {
   @javax.annotation.Nullable
   @ApiModelProperty(value = "")
 
-  public Long getUniqueLogicalSize() {
+  public Double getUniqueLogicalSize() {
     return uniqueLogicalSize;
   }
 
 
-  public void setUniqueLogicalSize(Long uniqueLogicalSize) {
+  public void setUniqueLogicalSize(Double uniqueLogicalSize) {
     this.uniqueLogicalSize = uniqueLogicalSize;
   }
 
 
-  public VmVolumeWhereInput uniqueLogicalSizeGt(Long uniqueLogicalSizeGt) {
+  public VmVolumeWhereInput uniqueLogicalSizeGt(Double uniqueLogicalSizeGt) {
     
     this.uniqueLogicalSizeGt = uniqueLogicalSizeGt;
     return this;
@@ -3759,17 +3759,17 @@ public class VmVolumeWhereInput {
   @javax.annotation.Nullable
   @ApiModelProperty(value = "")
 
-  public Long getUniqueLogicalSizeGt() {
+  public Double getUniqueLogicalSizeGt() {
     return uniqueLogicalSizeGt;
   }
 
 
-  public void setUniqueLogicalSizeGt(Long uniqueLogicalSizeGt) {
+  public void setUniqueLogicalSizeGt(Double uniqueLogicalSizeGt) {
     this.uniqueLogicalSizeGt = uniqueLogicalSizeGt;
   }
 
 
-  public VmVolumeWhereInput uniqueLogicalSizeGte(Long uniqueLogicalSizeGte) {
+  public VmVolumeWhereInput uniqueLogicalSizeGte(Double uniqueLogicalSizeGte) {
     
     this.uniqueLogicalSizeGte = uniqueLogicalSizeGte;
     return this;
@@ -3782,25 +3782,25 @@ public class VmVolumeWhereInput {
   @javax.annotation.Nullable
   @ApiModelProperty(value = "")
 
-  public Long getUniqueLogicalSizeGte() {
+  public Double getUniqueLogicalSizeGte() {
     return uniqueLogicalSizeGte;
   }
 
 
-  public void setUniqueLogicalSizeGte(Long uniqueLogicalSizeGte) {
+  public void setUniqueLogicalSizeGte(Double uniqueLogicalSizeGte) {
     this.uniqueLogicalSizeGte = uniqueLogicalSizeGte;
   }
 
 
-  public VmVolumeWhereInput uniqueLogicalSizeIn(List<Long> uniqueLogicalSizeIn) {
+  public VmVolumeWhereInput uniqueLogicalSizeIn(List<Double> uniqueLogicalSizeIn) {
     
     this.uniqueLogicalSizeIn = uniqueLogicalSizeIn;
     return this;
   }
 
-  public VmVolumeWhereInput addUniqueLogicalSizeInItem(Long uniqueLogicalSizeInItem) {
+  public VmVolumeWhereInput addUniqueLogicalSizeInItem(Double uniqueLogicalSizeInItem) {
     if (this.uniqueLogicalSizeIn == null) {
-      this.uniqueLogicalSizeIn = new ArrayList<Long>();
+      this.uniqueLogicalSizeIn = new ArrayList<Double>();
     }
     this.uniqueLogicalSizeIn.add(uniqueLogicalSizeInItem);
     return this;
@@ -3813,17 +3813,17 @@ public class VmVolumeWhereInput {
   @javax.annotation.Nullable
   @ApiModelProperty(value = "")
 
-  public List<Long> getUniqueLogicalSizeIn() {
+  public List<Double> getUniqueLogicalSizeIn() {
     return uniqueLogicalSizeIn;
   }
 
 
-  public void setUniqueLogicalSizeIn(List<Long> uniqueLogicalSizeIn) {
+  public void setUniqueLogicalSizeIn(List<Double> uniqueLogicalSizeIn) {
     this.uniqueLogicalSizeIn = uniqueLogicalSizeIn;
   }
 
 
-  public VmVolumeWhereInput uniqueLogicalSizeLt(Long uniqueLogicalSizeLt) {
+  public VmVolumeWhereInput uniqueLogicalSizeLt(Double uniqueLogicalSizeLt) {
     
     this.uniqueLogicalSizeLt = uniqueLogicalSizeLt;
     return this;
@@ -3836,17 +3836,17 @@ public class VmVolumeWhereInput {
   @javax.annotation.Nullable
   @ApiModelProperty(value = "")
 
-  public Long getUniqueLogicalSizeLt() {
+  public Double getUniqueLogicalSizeLt() {
     return uniqueLogicalSizeLt;
   }
 
 
-  public void setUniqueLogicalSizeLt(Long uniqueLogicalSizeLt) {
+  public void setUniqueLogicalSizeLt(Double uniqueLogicalSizeLt) {
     this.uniqueLogicalSizeLt = uniqueLogicalSizeLt;
   }
 
 
-  public VmVolumeWhereInput uniqueLogicalSizeLte(Long uniqueLogicalSizeLte) {
+  public VmVolumeWhereInput uniqueLogicalSizeLte(Double uniqueLogicalSizeLte) {
     
     this.uniqueLogicalSizeLte = uniqueLogicalSizeLte;
     return this;
@@ -3859,17 +3859,17 @@ public class VmVolumeWhereInput {
   @javax.annotation.Nullable
   @ApiModelProperty(value = "")
 
-  public Long getUniqueLogicalSizeLte() {
+  public Double getUniqueLogicalSizeLte() {
     return uniqueLogicalSizeLte;
   }
 
 
-  public void setUniqueLogicalSizeLte(Long uniqueLogicalSizeLte) {
+  public void setUniqueLogicalSizeLte(Double uniqueLogicalSizeLte) {
     this.uniqueLogicalSizeLte = uniqueLogicalSizeLte;
   }
 
 
-  public VmVolumeWhereInput uniqueLogicalSizeNot(Long uniqueLogicalSizeNot) {
+  public VmVolumeWhereInput uniqueLogicalSizeNot(Double uniqueLogicalSizeNot) {
     
     this.uniqueLogicalSizeNot = uniqueLogicalSizeNot;
     return this;
@@ -3882,25 +3882,25 @@ public class VmVolumeWhereInput {
   @javax.annotation.Nullable
   @ApiModelProperty(value = "")
 
-  public Long getUniqueLogicalSizeNot() {
+  public Double getUniqueLogicalSizeNot() {
     return uniqueLogicalSizeNot;
   }
 
 
-  public void setUniqueLogicalSizeNot(Long uniqueLogicalSizeNot) {
+  public void setUniqueLogicalSizeNot(Double uniqueLogicalSizeNot) {
     this.uniqueLogicalSizeNot = uniqueLogicalSizeNot;
   }
 
 
-  public VmVolumeWhereInput uniqueLogicalSizeNotIn(List<Long> uniqueLogicalSizeNotIn) {
+  public VmVolumeWhereInput uniqueLogicalSizeNotIn(List<Double> uniqueLogicalSizeNotIn) {
     
     this.uniqueLogicalSizeNotIn = uniqueLogicalSizeNotIn;
     return this;
   }
 
-  public VmVolumeWhereInput addUniqueLogicalSizeNotInItem(Long uniqueLogicalSizeNotInItem) {
+  public VmVolumeWhereInput addUniqueLogicalSizeNotInItem(Double uniqueLogicalSizeNotInItem) {
     if (this.uniqueLogicalSizeNotIn == null) {
-      this.uniqueLogicalSizeNotIn = new ArrayList<Long>();
+      this.uniqueLogicalSizeNotIn = new ArrayList<Double>();
     }
     this.uniqueLogicalSizeNotIn.add(uniqueLogicalSizeNotInItem);
     return this;
@@ -3913,12 +3913,12 @@ public class VmVolumeWhereInput {
   @javax.annotation.Nullable
   @ApiModelProperty(value = "")
 
-  public List<Long> getUniqueLogicalSizeNotIn() {
+  public List<Double> getUniqueLogicalSizeNotIn() {
     return uniqueLogicalSizeNotIn;
   }
 
 
-  public void setUniqueLogicalSizeNotIn(List<Long> uniqueLogicalSizeNotIn) {
+  public void setUniqueLogicalSizeNotIn(List<Double> uniqueLogicalSizeNotIn) {
     this.uniqueLogicalSizeNotIn = uniqueLogicalSizeNotIn;
   }
